+++ src/headSwingRetarget.js
@@ -1,5 +1,0 @@
-export {
-  CHAIN_AXES,
-  applyJointSwingTwist as applyHeadJointSwingTwist,
-  createChainItem as createHeadChainItem
-} from "./swingTwistRetarget.js";